+++ src/org/helioviewer/jhv/gui/dialogs/PreferencesDialog.java
@@ -173,12 +173,18 @@ public class PreferencesDialog extends StandardDialog implements ShowableDialog 
         normalizeRadius.addActionListener(e -> Settings.setProperty("display.normalize", Boolean.toString(normalizeRadius.isSelected())));
         settings.add(normalizeRadius, c);
 
-        c.gridx = 0;
+        c.gridx = 1;
         c.gridy = 5;
+        JCheckBox highResolution = new JCheckBox("Playback at high resolution", Boolean.parseBoolean(Settings.getProperty("display.highResolution")));
+        highResolution.addActionListener(e -> Settings.setProperty("display.highResolution", Boolean.toString(highResolution.isSelected())));
+        settings.add(highResolution, c);
+
+        c.gridx = 0;
+        c.gridy = 6;
         settings.add(new JLabel("Record video as:", JLabel.RIGHT), c);
 
         c.gridx = 1;
-        c.gridy = 5;
+        c.gridy = 6;
         JComboBox<VideoFormat> comboVideo = new JComboBox<>(VideoFormat.values());
         VideoFormat selected = VideoFormat.H264;
         try {
@@ -190,7 +196,7 @@ public class PreferencesDialog extends StandardDialog implements ShowableDialog 
         settings.add(comboVideo, c);
 
         c.gridx = 0;
-        c.gridy = 6;
+        c.gridy = 7;
         settings.add(new JLabel("Plugins:", JLabel.RIGHT), c);
 
         c.gridx = 1;